time/timecounter: Fix the lie that struct cyclecounter is const
BugLink: https://bugs.launchpad.net/bugs/2124107

In both the read callback for struct cyclecounter, and in struct
timecounter, struct cyclecounter is declared as a const pointer.

Unfortunatly, a number of users of this pointer treat it as a non-const
pointer as it is burried in a larger structure that is heavily modified by
the callback function when accessed.  This lie had been hidden by the fact
that container_of() "casts away" a const attribute of a pointer without any
compiler warning happening at all.

Fix this all up by removing the const attribute in the needed places so
that everyone can see that the structure really isn't const, but can,
and is, modified by the users of it.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Link: https://lore.kernel.org/all/2025070124-backyard-hurt-783a@gregkh
(backported from commit e78f70bad29c5ae1e1076698b690b15794e9b81e)
[bpoirier:
 Treewide change
 Conflicts:
 - drivers/net/ethernet/wangxun/libwx/wx_ptp.c
 	Due to the absence of
 	06e75161b9d4 net: wangxun: Add support for PTP clock (v6.15-rc1)
 	-> Skip this hunk]
Signed-off-by: Benjamin Poirier <bpoirier@nvidia.com>
Acked-by: Valentine Fatiev <valentinef@nvidia.com>
Acked-by: Amir Tzin <amirtz@nvidia.com>
Tested-by: Rola Dabbah <rdabbah@nvidia.com>
Tested-by: Shmuel Gutman <sgutman@nvidia.com>
Acked-by: Matthew R. Ochs <mochs@nvidia.com>
Acked-by: Carol L Soto <csoto@nvidia.com>
Acked-by: Kai-Heng Feng <kaihengf@nvidia.com>
Acked-by: Matt Ochs <mochs@nvidia.com>
Acked-by: Carol L Soto <csoto@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Abdur Rahman <abdur.rahman@canonical.com>
Acked-by: Ian May <ian.may@canonical.com>
Signed-off--by: Brad Figg <bfigg@nvidia.com>