panthor: save task pid and comm in panthor_group
BugLink: https://bugs.launchpad.net/bugs/2139960

[ Upstream commit 33b9cb6dcda2520600ac4fec725946af32c2e586 ]

We would like to report them on gpu errors.

We choose to save the info on panthor_group_create rather than on
panthor_open because, when the two differ, we are more interested in the
task that created the group.

Signed-off-by: Chia-I Wu <olvaffe@gmail.com>
Reviewed-by: Steven Price <steven.price@arm.com>
Signed-off-by: Steven Price <steven.price@arm.com>
Link: https://lore.kernel.org/r/20250718063816.1452123-3-olvaffe@gmail.com
Stable-dep-of: eec7e23d848d ("drm/panthor: Prevent potential UAF in group creation")
Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Manuel Diewald <manuel.diewald@canonical.com>
Signed-off-by: Edoardo Canepa <edoardo.canepa@canonical.com>