UBUNTU: [Config] move qcom interconnect/pinctrl/gcc as built-in for QCS8300
BugLink: https://bugs.launchpad.net/bugs/2134400

Move Qualcomm interconnect, pinctrl and global clock controllers from
QCS8300 to built-in to allow an early and functional boot on platforms
based on these devices.

Having these options as built-in removes the need for adding
clk_ignore_unused and pd_ignore_unused to the kernel command line
argument in order for the platform to boot.

Signed-off-by: Shivendra Pratap<shivendra.pratap@oss.qualcomm.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Timo Aaltonen <tjaalton@ubuntu.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>