NVIDIA: SAUCE: [Config] nvidia-6.11: drm: Update Kbuild to allow selecting CONFIG_DRM_CLIENT_SELECTION
BugLink: https://bugs.launchpad.net/bugs/2120973

Commit 8781125 backported a change to add the drm_client_setup()
function and the DRM_CLIENT_SELECTION configuration option, but didn't
packport any of the changes to make in-kernel DRM clients actually use
that function or select the config option. Thus, enabling the option
isn't possible and drm_client_setup() is always compiled to a stub
function that does nothing.

At compile time, nvidia-drm checks for the presence of
drm_client_setup() and calls that instead of drm_fbdev_ttm_setup() if
it's loaded with the modeset=1 and fbdev=1 parameters, which results in
a blank screen if CONFIG_DRM_CLIENT_SELECTION=n.

Update the option to make it selectable.

Fixes: 8781125 ("drm: Add client-agnostic setup helper")
Signed-off-by: Aaron Plattner <aplattner@nvidia.com>
Signed-off-by: Jamie Nguyen <jamien@nvidia.com>
Acked-by: Carol L Soto <csoto@nvidia.com>
Acked-by: Jamie Nguyen <jamien@nvidia.com>
Acked-by: Nirmoy Das <nirmoyd@nvidia.com>
Acked-by: Noah Wager <noah.wager@canonical.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Signed-off--by: Brad Figg <bfigg@nvidia.com>